+++ jdaviz/configs/specviz2d/plugins/spectral_extraction/tests/test_spectral_extraction.py
@@ -244,7 +244,7 @@ def test_horne_extract_self_profile(specviz2d_helper):
                           flux=spec2d*u.Jy,
                           uncertainty=VarianceUncertainty(spec2dvar*u.Jy*u.Jy))
 
-    specviz2d_helper.load(objectspec)
+    specviz2d_helper.load(objectspec, format='2D Spectrum')
     pext = specviz2d_helper.plugins['2D Spectral Extraction']._obj
 
     trace_fit = tracing.FitTrace(objectspec,
@@ -301,7 +301,7 @@ def test_horne_extract_self_profile(specviz2d_helper):
 
 
 def test_spectral_extraction_flux_unit_conversions(specviz2d_helper, mos_spectrum2d):
-    specviz2d_helper.load(mos_spectrum2d)
+    specviz2d_helper.load(mos_spectrum2d, format='2D Spectrum')
 
     uc = specviz2d_helper.plugins["Unit Conversion"]
     pext = specviz2d_helper.plugins['2D Spectral Extraction']