+++ jdaviz/core/loaders/importers/__init__.py
@@ -1,9 +1,8 @@
 from .importer import *  # noqa
 from .spectrum_common import *  # noqa
-from .spectrum1d import *  # noqa
-from .spectrum2d import *  # noqa
 from .spectrum3d import *  # noqa
-from .spectrum_list import *  # noqa
+from .spectrum2d import *  # noqa
+from .spectrum1d import *  # noqa
 from .ramp import *  # noqa
 from .ramp_integration import *  # noqa
 from .image import *  # noqa