+++ jdaviz/core/loaders/tests/test_loaders.py
@@ -53,9 +53,9 @@ def test_resolver_matching(specviz_helper):
 
     res_sp = find_matching_resolver(specviz_helper.app, sp)
     assert res_sp._obj._registry_label == 'object'
-    assert res_sp.format == '1D Spectrum'
+    assert '1D Spectrum' in res_sp.format.choices
 
-    specviz_helper._load(sp)
+    specviz_helper.load(sp)
     assert len(specviz_helper.app.data_collection) == 1
 
 
@@ -94,6 +94,7 @@ def test_trace_importer(specviz2d_helper, spectrum2d):
 def test_spectrum2d_viewer_options(deconfigged_helper, spectrum2d):
     ldr = deconfigged_helper.loaders['object']
     ldr.object = spectrum2d
+    ldr.format = '2D Spectrum'
 
     assert ldr.importer.viewer.create_new == '2D Spectrum'
     assert ldr.importer.viewer.new_label == '2D Spectrum'
@@ -158,7 +159,7 @@ def test_jwst_wfss_bsub(deconfigged_helper):
     ldr.cache = True
     ldr.url = uri
 
-    assert ldr.format == '2D Spectrum'
+    ldr.format = '2D Spectrum'  # may also be 'Image' depending on importer registry order
 
     ldr.load()
 
@@ -176,7 +177,7 @@ def test_fits_spectrum_list_L3_wfss(deconfigged_helper):
 
     # ldr = deconfigged_helper.loaders['file']
     # ldr.filepath = './jdaviz/notebooks/WFSS_fits/jw01076-o103_t0000_nircam_f356w-grismr_x1d.fits'  # noqa
-    ldr.format = '1D Spectrum List'
+    ldr.format = '1D Spectrum'
 
     # 1_117 is completely masked
     sources_obj = ldr.importer.sources
@@ -214,11 +215,11 @@ def test_resolver_url(deconfigged_helper):
     assert loader._obj.url_scheme == 's3'
     assert len(loader.format.choices) > 0
 
-    # https valid input
+    # https valid input (2D Spectrum)
     loader.url = 'https://stsci.box.com/shared/static/exnkul627fcuhy5akf2gswytud5tazmw.fits'  # noqa
 
     # may change with future importers
-    assert len(loader.format.choices) == 4
+    assert len(loader.format.choices) == 3
     assert loader.format.selected == 'Image'  # default may change with future importers
 
     # test target filtering
@@ -227,12 +228,12 @@ def test_resolver_url(deconfigged_helper):
     loader.target = '1D Spectrum'
 
     # may change with future importers
-    assert len(loader.format.choices) == 2
-    assert loader.format == '1D Spectrum List'  # default may change with future importers
-    assert loader.importer.data_label == 'exnkul627fcuhy5akf2gswytud5tazmw'  # noqa
+    assert len(loader.format.choices) == 1
+    assert loader.format == '1D Spectrum'  # default may change with future importers
+    assert loader.importer.data_label == 'exnkul627fcuhy5akf2gswytud5tazmw_index-0'  # noqa
 
     loader.target = 'Any'
-    assert len(loader.format.choices) == 4
+    assert len(loader.format.choices) == 3
     loader.format = '2D Spectrum'
     assert loader.importer.data_label == 'exnkul627fcuhy5akf2gswytud5tazmw'  # noqa